fixed admin realtime delay that was caused by useeffect and setstate being use together
